Modified email sending to use a function located in utils folder, this uses the same function as the branch Backend/110/registrationProblemHandler, added this now so that it is not forgotten later.
